JDK-8347270
